+++ engine.py
@@ -1,4 +1,4 @@
-from argelia_scraper.cli import main
+from uif_scraper.cli import main
 
 if __name__ == "__main__":
     main()